Split up transformation into index adding and actual transformation
