+++ src/audit-log-api/src/handlers/createAuditLogEvent.ts
@@ -6,7 +6,7 @@ import createAuditLogLookupDynamoDbConfig from "../createAuditLogLookupDynamoDbC
 import { AuditLogDynamoGateway, AwsAuditLogLookupDynamoGateway } from "../gateways/dynamo"
 import { CreateAuditLogEventUseCase, parseCreateAuditLogEventRequest, validateCreateAuditLogEvent } from "../use-cases"
 import StoreValuesInLookupTableUseCase from "../use-cases/StoreValuesInLookupTableUseCase"
-import { createJsonApiResult, transformAuditLogEvent } from "../utils"
+import { addAuditLogEventIndices, createJsonApiResult, transformAuditLogEvent } from "../utils"
 
 const auditLogConfig = createAuditLogDynamoDbConfig()
 const auditLogLookupConfig = createAuditLogLookupDynamoDbConfig()
@@ -35,7 +35,8 @@ export default async function createAuditLogEvent(event: APIGatewayProxyEvent): 
   }
 
   const transformedAuditLogEvent = transformAuditLogEvent(auditLogEvent)
-  const result = await createAuditLogEventUseCase.create(request.messageId, transformedAuditLogEvent)
+  const indexedAuditLogEvent = addAuditLogEventIndices(transformedAuditLogEvent)
+  const result = await createAuditLogEventUseCase.create(request.messageId, indexedAuditLogEvent)
 
   if (result.resultType === "notFound") {
     return createJsonApiResult({
