+++ src/audit-log-api/src/handlers/createAuditLogEvents.ts
@@ -10,7 +10,7 @@ import {
   validateCreateAuditLogEvents
 } from "../use-cases"
 import StoreValuesInLookupTableUseCase from "../use-cases/StoreValuesInLookupTableUseCase"
-import { createJsonApiResult, transformAuditLogEvent } from "../utils"
+import { addAuditLogEventIndices, createJsonApiResult, transformAuditLogEvent } from "../utils"
 
 const auditLogConfig = createAuditLogDynamoDbConfig()
 const auditLogLookupConfig = createAuditLogLookupDynamoDbConfig()
@@ -42,7 +42,8 @@ export default async function createAuditLogEvents(event: APIGatewayProxyEvent):
 
   const auditLogEvents = eventValidationResults.map((result) => result.auditLogEvent)
   const transformedAuditLogEvents = auditLogEvents.map(transformAuditLogEvent)
-  const result = await createAuditLogEventUseCase.create(request.messageId, transformedAuditLogEvents)
+  const indexedAuditLogEvents = transformedAuditLogEvents.map(addAuditLogEventIndices)
+  const result = await createAuditLogEventUseCase.create(request.messageId, indexedAuditLogEvents)
 
   if (result.resultType === "notFound") {
     return createJsonApiResult({
