+++ src/audit-log-api/src/utils/index.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/prefer-default-export
+export { default as addAuditLogEventIndices } from "./addAuditLogEventIndices"
 export { default as createJsonApiResult } from "./createJsonApiResult"
 export { default as isIsoDate } from "./isIsoDate"
 export { default as shouldFetchLargeObjects } from "./shouldFetchLargeObjects"
 export { default as transformAuditLogEvent } from "./transformAuditLogEvent"
-export { default as transformAuditLogEventAttributes } from "./transformAuditLogEventAttributes"
 export { default as validateAuditLogEvent } from "./validateAuditLogEvent"
