+++ src/audit-log-api/src/utils/transformAuditLogEvent.ts
@@ -1,15 +1,19 @@
 import type { AuditLogEvent } from "shared-types"
-import shouldLogForAutomationReport from "../gateways/dynamo/AuditLogDynamoGateway/shouldLogForAutomationReport"
-import shouldLogForTopExceptionsReport from "../gateways/dynamo/AuditLogDynamoGateway/shouldLogForTopExceptionsReport"
-import transformAuditLogEventAttributes from "./transformAuditLogEventAttributes"
 
 const transformAuditLogEvent = (event: AuditLogEvent): AuditLogEvent => {
-  const transformedEvent = transformAuditLogEventAttributes(event)
+  if (event.attributes.eventCode && typeof event.attributes.eventCode === "string") {
+    event.eventCode = event.attributes.eventCode
+    delete event.attributes.eventCode
+  }
 
-  transformedEvent._automationReport = shouldLogForAutomationReport(transformedEvent)
-  transformedEvent._topExceptionsReport = shouldLogForTopExceptionsReport(transformedEvent)
+  const user = event.attributes.user ?? event.attributes.User
+  if (typeof user === "string") {
+    event.user = user
+    delete event.attributes.user
+    delete event.attributes.User
+  }
 
-  return transformedEvent
+  return event
 }
 
 export default transformAuditLogEvent
